Implement supervision in the frequency server
High-level system description:
- frequency_supervisor `spawn_link`s and supervises the frequency server
- frequency server `monitor`s clients to deallocate frequencies if client fail
- clients `link` to frequency_supervisor (this way they'll exit if the frequency_supervisor exits)

State after restarting the frequency server:
If a client calls `deallocate` after server is restarted, the frequency server reply with an error and the client can ask the frequency server to restore its frequency. The frequency server tries to allocate the same frequency but if it's already allocated it assigns a new one if available; if frequencies are exhausted the client is notified with an error.