Refactor state handling
This comes after peer-reviewing a solution: this dude came up with a very simple pattern to persist state that is message passing. The frequency server sends a msg to its supervisor (which it's guaranteed to survive longer) and if it's restarted it'll get the old state as a parameter of its init function. If the supervisor dies than everything goes rogue but this is how it should be.